% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes_to_spatial_polygons.R
\name{boxes_to_spatial_polygons}
\alias{boxes_to_spatial_polygons}
\title{Convert xml crown annotations into projected bounding boxes.
In order to plot the annotations, they need to be projected and overlayed on RGB boxes}
\usage{
boxes_to_spatial_polygons(boxes, raster_object, project_boxes = TRUE)
}
\arguments{
\item{boxes}{A data frame with xmin, xmax, ymin, ymax columns. Each row is a crown bounding box.}

\item{raster_object}{A RGB raster to overlay annotations}

\item{project_boxes}{Whether boxes need to be projected to utm from image coordinates (origin 0,0 top left)}
}
\value{
SpatialPolygons object of annotations
}
\description{
Convert xml crown annotations into projected bounding boxes.
In order to plot the annotations, they need to be projected and overlayed on RGB boxes
}
\examples{
\donttest{
xml<-get_data("SJER_052","annotations")
annotations<-xml_parse(xml)
rgb_path<-get_data("SJER_052","rgb")
rgb<-raster::stack(rgb_path)
ground_truth <- boxes_to_spatial_polygons(annotations,rgb)
plot(ground_truth)
}s
}
