% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_precision_recall.R
\name{compute_precision_recall}
\alias{compute_precision_recall}
\title{Compute precision and recall statistics between predicted tree boxes and ground truth data}
\usage{
compute_precision_recall(
  ground_truth,
  predictions,
  threshold = 0.4,
  summarize = TRUE
)
}
\arguments{
\item{ground_truth}{A SpatialPolygonDataFrame of ground truth polygons.}

\item{predictions}{A SpatialPolygonDataFrame of prediction polygons.}

\item{threshold}{The intersection-over-union threshold for a prediction overlap with a ground truth to be considered a true positive.}

\item{summarize}{Logical. If true, return the precision and recall for this dataset, if false, return a data frame of matched crowns and the IoU overlap with ground truth.}
}
\value{
The recall and precision scores for the plot.
}
\description{
Compute precision and recall statistics between predicted tree boxes and ground truth data
}
\details{
The numeric identity of the crown is stored in a column named crown_id. An error will be raised if this column does not exist.
}
