% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowns.R
\docType{data}
\name{crowns}
\alias{crowns}
\title{Field-annotated Crowns}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 564 rows and 7 columns.}
\usage{
crowns
}
\description{
These are ‘field-annotated crowns’ in which an observer annotates the remote-sensing image on a tablet while standing in the field.
From Ordway Swisher Biological Station, Florida and Mountain Lake Biological Station
\itemize{
  \item indvdID Individual crown identifier
  \item plotID NEON plot name. This name overlaps with the sensor data. For example, the rgb data for SJER_052 is evaluation/rgb/SJER_052.tif
  \item siteID NEON site abbreviation see https://www.neonscience.org/field-sites/field-sites-map
  \item utmZone geospatial projection in utm wgs84
  \item pltNrth utm northing coordinate of centroid
  \item pltEstn utm easting coordinate of centroid
  \item geometry sf geometry polygon
}
}
\keyword{datasets}
