% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{unzip_download}
\title{Download sensor and annotation data for the NeonTreeEvaluation Benchmark}
\usage{
download(training = FALSE, savedir = NULL, force = F)

unzip_download(destination)
}
\arguments{
\item{training}{Download training data? See details}

\item{savedir}{Optional directory to save data in a new location. Defaults to package contents. Warning: Functions are designed for this location, only use this argument if you do not intend to run evaluation data.}
}
\description{
The NeonTreeEvaluation benchmark consists of two parts, 1) package code to run evaluation workflows, 2) evaluation data. Evaluation data is ~ 2GB in size and will be downloaded to package contents. See
The training data is a large set of training tiles (>5GB). Training tiles are geographically seperate from the evaluation data.
}
\examples{
 \donttest{
 download()
 list_rgb()
 }
}
\references{
Juniper Simonis, & Ethan White. (2019, July 11). weecology/portalcasting: hookup to zenodo (Version v0.8.0-1). Zenodo. http://doi.org/10.5281/zenodo.3332974

Weinstein, Ben G., et al. "Cross-site learning in deep learning RGB tree crown detection." Ecological Informatics 56 (2020): 101061.
}
\author{
The core zenodo download function was written by Juniper Simonis in Weecology/portalcasting (Simonis and White 2019). Adapted by Ben Weinstein.
}
