% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_field_crowns.R
\name{evaluate_field_crowns}
\alias{evaluate_field_crowns}
\title{Compute intersection-over-union scores from the field collected crowns}
\usage{
evaluate_field_crowns(submission, summarize = T, project = FALSE)
}
\arguments{
\item{submission}{A five column dataframe in the order plot_name, xmin, xmax, ymin, ymax. Each row is a predicted bounding box.}
}
\value{
A data frame with the crown ID matched to the prediction ID.
}
\description{
\code{evaluate_field_crowns} implements the matching and scoring algorithm on crowns that were drawn on a tablet while physically standing in the field.
}
\details{
For details on the protocol for field-based crown delineation see Graves et al. (2018). see \code{field_crowns}
}
\examples{

}
\references{
Graves S, Gearhart J, Caughlin TT, Bohlman S. 2018. A digital mapping method for linking high-resolution remote sensing images to individual tree crowns. PeerJ Preprints 6:e27182v1 https://doi.org/10.7287/peerj.preprints.27182v1
}
