% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_plot.R
\name{evaluate_plot}
\alias{evaluate_plot}
\title{Compute evaluation statistics for one plot}
\usage{
evaluate_plot(submission, show = TRUE, project_boxes = TRUE)
}
\arguments{
\item{submission}{A five column dataframe in the order plot_name, xmin, xmax, ymin, ymax. Each row is a predicted bounding box.}

\item{show}{Logical. Plot the overlayed annotations for each plot?}

\item{project_boxes}{Logical. Should the boxes be projected into utm coordinates? This is needed if the box coordinates are given from the image origin (top left is 0,0)}

\item{compute_PR}{Logical. Should the average precision and recall be computed?}
}
\value{
If compute_PR=T, the recall and precision scores for the plot, if False, the intersection-over-union scores for each prediction.
}
\description{
Compute evaluation statistics for one plot
}
