% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{data}
\name{field}
\alias{field}
\title{NEON Woody Vegetation Structure}
\format{
An object of class \code{data.frame} with 37776 rows and 67 columns.
}
\usage{
field
}
\description{
From https://data.neonscience.org/data-products/DP1.10098.001
}
\details{
This data product contains the quality-controlled, native sampling resolution data from in-situ measurements of live and standing dead woody individuals and shrub groups, from all terrestrial NEON sites with qualifying woody vegetation. The exact measurements collected per individual depend on growth form, and these measurements are focused on enabling biomass and productivity estimation, estimation of shrub volume and biomass, and calibration / validation of multiple NEON airborne remote-sensing data products. In general, comparatively large individuals that are visible to remote-sensing instruments are mapped, tagged and measured, and other smaller individuals are tagged and measured but not mapped. Smaller individuals may be subsampled according to a nested subplot approach in order to standardize the per plot sampling effort. Structure and mapping data are reported per individual per plot; sampling metadata, such as per growth form sampling area, are reported per plot. For additional details, see the user guide, protocols, and science design listed in the Documentation section in this data product's details webpage.
}
\keyword{datasets}
