% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_plot.R, R/image_crowns.R
\name{image_crowns}
\alias{image_crowns}
\alias{image_crowns}
\title{Compute evaluation statistics for one plot}
\usage{
image_crowns(submission, show = TRUE, project_boxes = TRUE)

image_crowns(submission, show = TRUE, project_boxes = TRUE)
}
\arguments{
\item{submission}{A five column dataframe in the order plot_name, xmin, xmax, ymin, ymax. Each row is a predicted bounding box.}

\item{show}{Logical. Plot the overlayed annotations for each plot?}

\item{project_boxes}{Logical. Should the boxes be projected into utm coordinates? This is needed if the box coordinates are given from the image origin (top left is 0,0)}

\item{compute_PR}{Logical. Should the average precision and recall be computed?}

\item{submission}{The format of the submission is a csv with 5 columns: plot_name, xmin, ymin, xmax, ymax  follows
Each row contains information for one predicted bounding box.
The plot column should be named the same as the files in the dataset (e.g. SJER_021), not the path to the file.}

\item{show}{Logical. Plot the overlayed annotations for each plot?}

\item{project_boxes}{Logical. Should the boxes be projected into utm coordinates? This is needed if the box coordinates are given from the image origin (top left is 0,0)}

\item{compute_PR}{Logical. Should the average precision and recall be computed?}
}
\value{
If compute_PR=T, the recall and precision scores for the plot, if False, the intersection-over-union scores for each prediction.

If compute_PR=T, the recall and precision scores for the plot, if False, the intersection-over-union scores for each prediction.
}
\description{
Compute evaluation statistics for one plot

Compute evaluation statistics for one plot of image-annotated crowns
}
