% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_annotations.R
\name{list_annotations}
\alias{list_annotations}
\title{List paths to annotations}
\usage{
list_annotations()
}
\description{
This function looks into package contents for ground truth annotations for the image-annotated crowns. See download() to download full set of data from zenodo.
}
