% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chm.R
\name{list_chm}
\alias{list_chm}
\title{List paths to canopy height files}
\usage{
list_chm()
}
\description{
This function looks into package contents (system.file("extdata")). See download() to download full set of sensor data from zenodo.
}
