% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_field_crown.R
\name{load_field_crown}
\alias{load_field_crown}
\title{Load and overlay ground truth annotations for single plot evaluation}
\usage{
load_field_crown(plot_name, show = TRUE)
}
\arguments{
\item{plot_name}{The name of plot as given by the filename (e.g "SJER_021.tif" -> SJER_021).}

\item{show}{Logical. Whether to plot  the ground truth data overlayed on the RGB image.}
}
\value{
A SpatialPolygonsDataFrame of ground truth boxes.
}
\description{
load_ground_truth is a wrapper function to get a plot annotation from file, project into geographic coordinates and potentially overlay on RGB data
}
