% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR.R
\name{PR}
\alias{PR}
\title{Precision and recall for a set of prediction results}
\usage{
PR(results, threshold)
}
\arguments{
\item{results}{A dataframe returned from \code{evaluate_plot}}
}
\value{
Summary recall and precision
}
\description{
\code{PR} is an internal function not called by the user for summary statistics
}
