% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_detection.R
\name{stem_detection}
\alias{stem_detection}
\title{Evaluate predictions against NEON field data}
\usage{
stem_detection(submission, project_boxes = FALSE, show = T,
  field_data = NA)
}
\arguments{
\item{submission}{A five column dataframe in the order plot_name, xmin, xmax, ymin, ymax. Each row is a predicted bounding box.}

\item{project_boxes}{Logical. Should the boxes be projected into utm coordinates? This is needed if the box coordinates are given from the image origin (top left is 0,0)}

\item{show}{Logical. Plot the overlayed annotations for each plot?}

\item{field_data}{a csv file with spatial coordinates "easting" and "northing" corresponding to utm locations of each point}
}
\value{
The average stem recall scores for each image in the submission
}
\description{
Evaluate predictions against NEON field data
}
\details{
For each plot in the submission, this function will check if there are field collected stem data and score whether each stem is within a predicted tree bounding box
}
