% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Summary statistics for benchmark evaluation}
\usage{
summary_statistics(results, method = "all", plot_result = T)
}
\arguments{
\item{results}{A dataframe returned from \code{evaluate_plot}}

\item{method}{"all" for averaging recall across images, or "site" to average per geographic site}

\item{plot_result}{generate a ggplot of results}
}
\value{
Summary recall and precision
}
\description{
\code{summary_statistics} returns a helpful summary of a set of plot evaluations
}
