% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_training.R
\name{zenodo_url}
\alias{zenodo_url}
\title{Download training data tiles for the NeonTreeEvaluation Benchmark}
\usage{
zenodo_url(concept_rec_id = 3459803, rec_version = "latest",
  rec_id = NULL, arg_checks = TRUE)
}
\description{
The benchmark dataset contains both evaluation and training data for future use. While the annotations and evaluation data are modest in size, the training data is over 20GB and cannot be included directly in the package.
To download the data, the download_data funtion pulls the training tiles from the Zenodo archive and places them alongside the evalution data by default
}
