% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_field_crowns.R
\name{evaluate_field_crowns}
\alias{evaluate_field_crowns}
\title{Compute evaluation metrics for the field-collected crowns}
\usage{
evaluate_field_crowns(submission, summarize = TRUE, show = TRUE,
  project = FALSE)
}
\arguments{
\item{submission}{The format of the submission is a csv with 5 columns: plot_name, xmin, ymin, xmax, ymax  follows
Each row contains information for one predicted bounding box.
The plot column should be named the same as the files in the dataset (e.g. SJER_021), not the path to the file.}

\item{summarize}{Compute summary statistics for crown recall}

\item{show}{Logical. Plot the overlayed annotations for each plot?}

\item{project}{Whether to project the supplied bounding box coordinates from image coordinates into geographic coordinates (utm QGS84). This is needed for computing recall scores.}
}
\value{
If summarize is True, a set of summary measures from \code{summary_statistics} for the overall score, the entire site score, and the per-plot score.
If False, a dataframe with the intersection-over-union scores for each prediction.
}
\description{
\code{evaluate_field_crowns} implements the matching and scoring algorithm on crowns that were drawn on a tablet while physically standing in the field.
}
\details{
For details on the protocol for field-based crown delineation see Graves et al. (2018). see \code{field_crowns}
Not all evaluation data are available for all plots. This function will look for matching plot name and ignore other plots.
}
\examples{
data("submission")
df <- submission \%>\% dplyr::filter(plot_name=="OSBS_95_competition")
results <- evaluate_field_crowns(submission = df,project = FALSE, summarize = TRUE)
}
\references{
Graves S, Gearhart J, Caughlin TT, Bohlman S. 2018. A digital mapping method for linking high-resolution remote sensing images to individual tree crowns. PeerJ Preprints 6:e27182v1 https://doi.org/10.7287/peerj.preprints.27182v1
}
