% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_field_stems.R
\name{evaluate_field_stems}
\alias{evaluate_field_stems}
\title{Compute evaluation scores from the field collected stem data}
\usage{
evaluate_field_stems(submission, project = TRUE, show = T)
}
\value{
A data frame with the number of crown predictions in the sumbitted predictions and the field sampled plots.
}
\description{
\code{evaluate_field_stems} implements the matching and scoring algorithm on field stems that were collected by the NEON Woody Vegetation structure data.
}
\details{
The following filters were applied to the raw NEON field data (ID) after download. An overstory reference tree must have:
\itemize{
\item Valid spatial coordinates
\item A unique height measurement per sampling period. Species double recorded but with different heights were discarded
\item Sampled in more than one year to verify height measurement
\item Changes in between year field heights of less than 6m
\item A minimum height of 3m to match the threshold in the remote sensing workflow.
\item Be at least within 5m of the canopy as measured by the LiDAR height model extracted at the stem location. The was used to prevent matching with understory trees in the event that overstory trees were eliminated due to failing in one of the above conditions, or not sampled by NEON.
}
}
