% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_stem.R
\name{evaluate_stem}
\alias{evaluate_stem}
\title{Evaluate predictions against NEON field data}
\usage{
evaluate_stem(plot_prediction, stem_dat, project_boxes = FALSE, show = T)
}
\arguments{
\item{stem_dat}{Object. A sf object with utm coordinates}

\item{project_boxes}{Logical. Should the boxes be projected into utm coordinates? This is needed if the box coordinates are given from the image origin (top left is 0,0)}

\item{show}{Logical. Plot the overlayed annotations for each plot?}

\item{submission}{A five column dataframe in the order plot_name, xmin, xmax, ymin, ymax. Each row is a predicted bounding box.}
}
\value{
The recall scores for each image
}
\description{
Evaluate predictions against NEON field data
}
\details{
For each plot in the submission, this function will check if there are field collected stem data and score whether each stem is within a predicted tree bounding box
}
