% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve benchmark data
\code{get_data} is a set of utility functions for finding the path of benchmark data on disk}
\usage{
get_data(plot_name, sensor)
}
\arguments{
\item{plot_name}{A plot name}

\item{sensor}{Which plot object should be returned: "rgb","lidar","hyperspectral","annotations"}
}
\value{
path to object on disk
}
\description{
Retrieve benchmark data
\code{get_data} is a set of utility functions for finding the path of benchmark data on disk
}
