% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_annotations.R
\name{list_annotations}
\alias{list_annotations}
\title{Find plot names with annotations}
\usage{
list_annotations()
}
\value{
A vector of plot names with available annotations
}
\description{
\code{list_annotations} returns the plot names with available annotations in this release of the NeonTreeEvaluation benchmark. There are many plots for which their is sensor data, but no annotations.
}
