% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_field_crowns.R
\name{list_field_crowns}
\alias{list_field_crowns}
\title{List images that overlap with field-annotation crowns}
\usage{
list_field_crowns()
}
\description{
Use this function to determine which sensor data overlaps with field collected crown data
}
